Added memory usage display onto StatusView.
Sorry I didn't fork the repo for this! I just had this change locally so I was thinking why not commit(since this is a plugin and stuff). Maybe a toggle or something can be added too, although I prefer it this way for less lag. Hope this makes it.