Move rencache global data into a struct
Now we put all the data needed by rencache into a struct and pass a pointer to
the struct to all the rencache functions to make it reentrant.

A global rencache struct is declared now in main.c so all the
operations will work in the same way except we always pass the
pointer to rencache struct.

This work is meant to transition into a system where the rencache
machinery can be used to render and update a part of the window.

Currently the rencache performs all the drawing operations into the
implicit window's surface. Future work may change this so that rencache
can write into an arbitrary surface.