WIP: working on the draw_image rencache command
Work in progress, do not compile and it is largely incomplete.

The idea is to implement a new command into the rencache, DRAW_IMAGE.
The command should blit into the rendering surface a rectangular
portion of a given image.

We should have a "repository" of images to be used for this operations.
Each image should be identified by an image ID. The image ID in turn
should have three components:

- the node the image belongs to (each node can potentially produce
  multiple images).
- an identifier of a specific rect withing a specific image of the node.
  For example
  for a DocView we may divide the document in pages and draw each
  page in a different image. We should in addition divide the image in
  rects so that an update in the image does not trigger the update
  of the whole image but only of a pert of it.
  The page and the rect within would provide a way to index
  the image.
- an identifier of the "revision" of the image. The idea is that to
  trigger an update upon a DRAW_IMAGE command the image_id has to
  change. If the image is always the same the "revision" number will
  track updates within an existing image.

In turn we should allow the views to render into an image and then
issue rencache commands to draw parts of the image (rects) using
the DRAW_IMAGE command. This part is not yet implemented.