Make `NagView` focus on correct `View` after close
When `NagView` has `force_focus`, any other `View` that tries to gain 
focus is rejected; we memorize the requesting `View`, so that we can 
focus on it when `NagView` closes.

An example of this is when we are switching project directories and we 
have an unsaved file. Before the `NagView` is able to close, we have 
already loaded the new project and asked for focus on a file. With this 
commit, the file is correctly focused when the `NagView` closes.