More accurate logic for node's remove view method
Provide an fix like:

https://github.com/lite-xl/lite-xl/pull/526

but use instead the is_primary_node property and the context
property.

We align the logic to the one used in the close_all_views method.

The approach using get_locked_size was the old approach using by
rxi's lite but it should no longer be used to determine if a node
if part of the application UI or part of the document's view.