Use better LCD filter weight for subpixel rendering
Based on:

https://freetype.org/freetype2/docs/reference/ft2-lcd_rendering.html

the coefficients should be normalized which was not the case. We use
the values previously used with the AGG library implementation. The
new coefficients are normalized and creates maybe less color fringes.

Ultimately it is difficult and subjective to say which coefficients are
better and the answer may depends on the vision of the person or on the
characteristics of the used monitor.