Check last used project before trying to use it
When starting the application try to uses the last recently used project
but doesn't check if the directory actually exists.

On windows that's okay because if it doesn't exist core.set_project_dir will
fail and the application will fall back to use the current working directory.

On linux on the other side it may cause an error with dmon because we attempt
to watch a directory that doesn't exist.

We now check that the directory actually exists before trying to use it.

This is an attempt to fix:

https://github.com/lite-xl/lite-xl/issues/774