add unicode support for windows with wchar
yes, you heard it right. We're using the worst of the worst
practices ISO C and Windows have to offer. Want to know why?
Because MS is a coporate infant that is unwilling to adopt change
early on and come to regret it later, but provides a "fix" that
doesn't work.

On Windows, you're supposed to have UTF8 support by
setlocale(LC_ALL, ".UTF-8"). This however will break EVERYTHING
that doesn't expect UTF-8 AND only works after Windows 10 2017 update.
It also (by all MS logic) isn't enabled by default.
Now, we can't expect our users to navigate through the mess that is
the Windows Language and Region setting and check a checkbox that has
BETA on it, right? Talking about catching on the trend.

This commit instead converts (back and forth) UTF-8 to UTF-16 in wchars.
There's a lot that can go wrong, and lots of code to account for.