Cache scaled canvas if `renderer.draw_surface` requires scaling
This avoids artifacts caused by rounding errors in clipping.

When we need the scaled version of a canvas, we check if we have already
one cached in the current frame cache.
If not we check if it was cached in the previous frame, and if so, we
copy it to the current frame cache.

If the scaled canvas is not found, we create one and push it to the
current frame cache.

We keep the cache of the previous frame to avoid recreating the same
scaled canvas every frame.